refactor: break up the Modal components
