+++ packages/@tinacms/react-modals/src/ModalProvider.tsx
@@ -17,11 +17,7 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { useCallback, useState, ReactNode } from 'react'
-import { createPortal } from 'react-dom'
-import styled, { StyledComponent } from 'styled-components'
-import { CloseIcon } from '@tinacms/icons'
-import { Button, TinaReset } from '@tinacms/styles'
+import { useCallback, useState } from 'react'
 
 interface Props {
   children: any
@@ -67,116 +63,3 @@ export function useModalContainer(): ModalContainerProps {
 
   return modalContainer
 }
-
-export const Modal = ({
-  ...props
-}: React.DetailedHTMLProps<
-  React.HTMLAttributes<HTMLDivElement>,
-  HTMLDivElement
->) => {
-  const { portalNode } = useModalContainer()
-
-  return portalNode && portalNode
-    ? (createPortal(
-        <TinaReset>
-          <ModalOverlay>
-            <div {...props} />
-          </ModalOverlay>
-        </TinaReset>,
-        portalNode
-      ) as any)
-    : null
-}
-
-export const ModalOverlay = styled.div`
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0, 0, 0, 0.5);
-  overflow: auto;
-  padding: 0;
-  z-index: var(--tina-z-index-3);
-`
-
-const ModalTitle = styled.h2`
-  font-size: var(--tina-font-size-4);
-  font-weight: var(--tina-font-weight-regular);
-  line-height: normal;
-  margin: 0;
-`
-
-const CloseButton = styled.div`
-  display: flex;
-  align-items: center;
-  fill: var(--tina-color-grey-5);
-  cursor: pointer;
-  transition: fill 85ms ease-out;
-  svg {
-    width: 24px;
-    height: auto;
-  }
-  &:hover {
-    fill: var(--tina-color-grey-8);
-  }
-`
-
-export interface ModalHeaderProps {
-  children: ReactNode
-  close?(): void
-}
-
-export const ModalHeader = styled(
-  ({ children, close, ...styleProps }: ModalHeaderProps) => {
-    return (
-      <div {...styleProps}>
-        <ModalTitle>{children}</ModalTitle>
-        {close && (
-          <CloseButton onClick={close}>
-            <CloseIcon />
-          </CloseButton>
-        )}
-      </div>
-    )
-  }
-)`
-  height: 56px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 var(--tina-padding-big) 0 var(--tina-padding-big);
-  border-bottom: 1px solid var(--tina-color-grey-3);
-  margin: 0;
-`
-
-export const ModalBody = styled.div<{ padded?: boolean }>`
-  padding: ${p => (p.padded ? 'var(--tina-padding-big)' : '0')};
-  margin: 0;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  min-height: 160px;
-
-  &:last-child {
-    border-radius: 0 0 5px 5px;
-  }
-`
-
-export const ModalActions: StyledComponent<'div', {}, {}> = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  border-radius: 0 0 5px 5px;
-  padding: 0 var(--tina-padding-big) var(--tina-padding-big)
-    var(--tina-padding-big);
-  ${Button} {
-    flex: 0 1 auto;
-    min-width: 128px;
-    margin: 0 var(--tina-padding-small) 0 0;
-    &:last-child {
-      margin-right: 0;
-    }
-  }
-`
